refactor: 공통되는 type을 별도의 interface로 분리 - #50
